server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
